// src/components/constructor/side-bar/script-block/widget/types.tsx
export enum InputTypeType {
    id = 'id',
    textarea = 'textarea',
    number = 'number',
    select = 'select',
    toggle = 'toggle',
    data = 'data',
    auto = 'auto',
    stage = 'stage',
    stages = 'stages',
    path = 'path',
    image = 'image',
    key = 'key',
    table = 'table',
    multiKey = 'multiKey',
    coords = 'coords',
    scene = 'scene',
    genericTable = 'genericTable',
    empty = 'empty',
    doubleNumber = 'doubleNumber',
}